添加全局异常处理

+++ xuecheng-content/xuecheng-content-model/src/main/java/cwb/content/model/domain/CourseMarket.java
@@ -28,12 +28,12 @@ public class CourseMarket implements Serializable {
     /**
      * 现价
      */
-    private Double price;
+    private Float price;
 
     /**
      * 原价
      */
-    private Double originalPrice;
+    private Float originalPrice;
 
     /**
      * 咨询qq
